+++ web/app/scripts/compiled-es5/services/calculate-percentage.js
@@ -19,15 +19,18 @@ var moduleProperties = exports.moduleProperties = {
 var _module = _libraries.angular.module(moduleProperties.moduleName, []);
 
 _module.service(moduleProperties.componentName, [function () {
-    this.countQuestionsTotal = function (form) {
-        var questions = form;
-        var totalQuestionCount = 0;
 
-        _libraries.angular.forEach(questions, function (question) {
-            totalQuestionCount = totalQuestionCount + question.fieldGroup.length;
+    this.getImportantQuestionKeys = function (form) {
+        var qKeys = [];
+        _libraries.angular.forEach(form, function (question) {
+            _libraries.angular.forEach(question.fieldGroup, function (field) {
+                if (field.templateOptions.fieldOptions.important === true) {
+                    qKeys.push(field.key);
+                }
+            });
         });
 
-        return totalQuestionCount;
+        return qKeys;
     };
 
     // answers is an object containing objects
@@ -38,15 +41,21 @@ _module.service(moduleProperties.componentName, [function () {
     //     }
     // }
     // answered is a number of type number
-    this.countNumberOfAnsweredQuestions = function (answers) {
+    this.countNumberOfAnsweredQuestions = function (answers, qKeys) {
         var answered = 0;
+
+        if (qKeys.length === 0) {
+            return answered;
+        }
 
         for (var groupKey in answers) {
             var answerGroup = answers[groupKey];
             for (var answerKey in answerGroup) {
                 var answer = answerGroup[answerKey];
-                if (_libraries.angular.isDefined(answer) && answer !== null) {
-                    answered++;
+                if (qKeys.indexOf(parseInt(answerKey, 10)) !== -1) {
+                    if (answer && answer !== null || answer === 0) {
+                        answered++;
+                    }
                 }
             }
         }