+++ web/app/scripts/compiled-es5/services/fieldsets-parser.js
@@ -36,6 +36,7 @@ _module.service(moduleProperties.componentName, [function () {
                                 // jscs:disable disallowQuotedKeysInObjects
                                 'help_text': field.help_text,
                                 required: field.required,
+                                important: field.important,
                                 label: field.name
                                 // jscs:enable disallowQuotedKeysInObjects
                             }