+++ web/tests/calculate-percentage-service-spec.js
@@ -2,16 +2,64 @@ describe('calculateCompletionPercentService', function () {
     const questionSet = {
         1: {
             fieldGroup: [
-                1,
-                2,
-                3
+                {
+                    key: 1,
+                    templateOptions: {
+                        fieldOptions: {
+                            important: false
+                        }
+                    }
+
+                },
+                {
+                    key: 2,
+                    templateOptions: {
+                        fieldOptions: {
+                            important: true
+                        }
+                    }
+
+                },
+                {
+                    key: 3,
+                    templateOptions: {
+                        fieldOptions: {
+                            important: true
+                        }
+                    }
+
+                }
             ]
         },
         2: {
             fieldGroup: [
-                1,
-                2,
-                3
+                {
+                    key: 1,
+                    templateOptions: {
+                        fieldOptions: {
+                            important: false
+                        }
+                    }
+
+                },
+                {
+                    key: 2,
+                    templateOptions: {
+                        fieldOptions: {
+                            important: true
+                        }
+                    }
+
+                },
+                {
+                    key: 3,
+                    templateOptions: {
+                        fieldOptions: {
+                            important: false
+                        }
+                    }
+
+                }
             ]
         }
     };
@@ -23,12 +71,12 @@ describe('calculateCompletionPercentService', function () {
             this.Parser = calculateCompletionPercent;
         });
 
-        this.totalQuestions = this.Parser.countQuestionsTotal(questionSet);
+        this.keys = this.Parser.getImportantQuestionKeys(questionSet);
     });
 
     describe('countQuestionsTotal function', function () {
         it('should return the total number of questions', function () {
-            expect(this.totalQuestions).toBe(6);
+            expect(this.keys.length).toBe(3);
         });
     });
 
@@ -40,7 +88,7 @@ describe('calculateCompletionPercentService', function () {
                     2: null
                 }
             };
-            const noQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(emptyAnswerSet);
+            const noQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(emptyAnswerSet, this.keys);
             expect(noQuestionsAnswered).toBe(0);
         });
 
@@ -50,18 +98,18 @@ describe('calculateCompletionPercentService', function () {
                     2: undefined
                 }
             };
-            const noQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(emptyAnswerSet);
+            const noQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(emptyAnswerSet, this.keys);
             expect(noQuestionsAnswered).toBe(0);
         });
 
-        it('should count empty string answers', function () {
+        it('should not count empty string answers', function () {
             const emptyAnswerSet = {
                 1: {
                     2: ''
                 }
             };
-            const noQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(emptyAnswerSet);
-            expect(noQuestionsAnswered).toBe(1);
+            const noQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(emptyAnswerSet, this.keys);
+            expect(noQuestionsAnswered).toBe(0);
         });
 
         it('should count the correct number of mixed questions answered', function () {
@@ -76,8 +124,8 @@ describe('calculateCompletionPercentService', function () {
                     6: ''
                 }
             };
-            const partialQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(partialAnswerSet);
-            expect(partialQuestionsAnswered).toBe(4);
+            const partialQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(partialAnswerSet, this.keys);
+            expect(partialQuestionsAnswered).toBe(2);
         });
 
         it('should count the correct number when all questions answered', function () {
@@ -93,8 +141,8 @@ describe('calculateCompletionPercentService', function () {
                     6: 5
                 }
             }
-            const allQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(allAnswerSet);
-            expect(allQuestionsAnswered).toBe(6);
+            const allQuestionsAnswered = this.Parser.countNumberOfAnsweredQuestions(allAnswerSet, this.keys);
+            expect(allQuestionsAnswered).toBe(2);
         });
     });
 
@@ -105,22 +153,22 @@ describe('calculateCompletionPercentService', function () {
         });
 
         it('should return the 0 percentage when none answered', function () {
-            this.percentage = this.Parser.calculatePercentageComplete(0, this.totalQuestions);
+            this.percentage = this.Parser.calculatePercentageComplete(0, this.keys.length);
             expect(this.percentage).toBe(0 + '%');
         });
 
-        it('should return 33% when 2/6 questions answered', function () {
-            this.percentage = this.Parser.calculatePercentageComplete(2, this.totalQuestions);
+        it('should return 33% when 2/3 questions answered', function () {
+            this.percentage = this.Parser.calculatePercentageComplete(1, this.keys.length);
             expect(this.percentage).toBe(33 + '%');
         });
 
-        it('should return 50% when 3/6 questions answered', function () {
-            this.percentage = this.Parser.calculatePercentageComplete(3, this.totalQuestions);
-            expect(this.percentage).toBe(50 + '%');
+        it('should return 67% when 3/3 questions answered', function () {
+            this.percentage = this.Parser.calculatePercentageComplete(2, this.keys.length);
+            expect(this.percentage).toBe(67 + '%');
         });
 
         it('should return the 100 percentage when all answered', function () {
-            this.percentage = this.Parser.calculatePercentageComplete(6, this.totalQuestions);
+            this.percentage = this.Parser.calculatePercentageComplete(3, this.keys.length);
             expect(this.percentage).toBe(100 + '%');
         });
     });